# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentCraft API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type WorkflowNodeType =
  (typeof WorkflowNodeType)[keyof typeof WorkflowNodeType];

export const WorkflowNodeType = {
  input: "input",
  ai_agent: "ai_agent",
  api_call: "api_call",
  condition: "condition",
  loop: "loop",
  output: "output",
} as const;

export type WorkflowNodeConfig = { [key: string]: unknown };

export type WorkflowNodePosition = {
  x: number;
  y: number;
};

export interface WorkflowNode {
  id: string;
  type: WorkflowNodeType;
  label: string;
  config?: WorkflowNodeConfig;
  position: WorkflowNodePosition;
}

export interface WorkflowEdge {
  id: string;
  source: string;
  target: string;
  label?: string;
}

export interface Workflow {
  id: number;
  name: string;
  description?: string;
  nodes: WorkflowNode[];
  edges: WorkflowEdge[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateWorkflowBody {
  name: string;
  description?: string;
  nodes: WorkflowNode[];
  edges: WorkflowEdge[];
}

export interface GenerateWorkflowBody {
  prompt: string;
}

export interface GeneratedWorkflow {
  name: string;
  description?: string;
  nodes: WorkflowNode[];
  edges: WorkflowEdge[];
}

export interface WorkflowExplanation {
  explanation: string;
  steps: string[];
}

export type NodeResultStatus =
  (typeof NodeResultStatus)[keyof typeof NodeResultStatus];

export const NodeResultStatus = {
  pending: "pending",
  running: "running",
  success: "success",
  failed: "failed",
  skipped: "skipped",
} as const;

export type NodeResultOutput = { [key: string]: unknown };

export interface NodeResult {
  nodeId: string;
  nodeType: string;
  label: string;
  status: NodeResultStatus;
  output?: NodeResultOutput;
  reasoning?: string;
  startedAt?: string;
  completedAt?: string;
  durationMs?: number;
}

export type ExecutionStatus =
  (typeof ExecutionStatus)[keyof typeof ExecutionStatus];

export const ExecutionStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
  cancelled: "cancelled",
} as const;

export interface Execution {
  id: number;
  workflowId: number;
  status: ExecutionStatus;
  input: string;
  createdAt: string;
  updatedAt: string;
}

export type ExecutionDetailStatus =
  (typeof ExecutionDetailStatus)[keyof typeof ExecutionDetailStatus];

export const ExecutionDetailStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
  cancelled: "cancelled",
} as const;

export interface ExecutionDetail {
  id: number;
  workflowId: number;
  status: ExecutionDetailStatus;
  input: string;
  finalOutput?: string;
  nodeResults: NodeResult[];
  agentLogs: string[];
  createdAt: string;
  updatedAt: string;
}

export interface StartExecutionBody {
  workflowId: number;
  input: string;
}

export type ListExecutionsParams = {
  workflowId?: number;
};