# lib/api-zod/src/generated/types/execution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionStatus } from "./executionStatus";

export interface Execution {
  id: number;
  workflowId: number;
  status: ExecutionStatus;
  input: string;
  createdAt: Date;
  updatedAt: Date;
}