# lib/api-zod/src/generated/types/executionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionDetailStatus } from "./executionDetailStatus";
import type { NodeResult } from "./nodeResult";

export interface ExecutionDetail {
  id: number;
  workflowId: number;
  status: ExecutionDetailStatus;
  input: string;
  finalOutput?: string;
  nodeResults: NodeResult[];
  agentLogs: string[];
  createdAt: Date;
  updatedAt: Date;
}