# lib/api-zod/src/generated/types/generatedWorkflow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowEdge } from "./workflowEdge";
import type { WorkflowNode } from "./workflowNode";

export interface GeneratedWorkflow {
  name: string;
  description?: string;
  nodes: WorkflowNode[];
  edges: WorkflowEdge[];
}