# lib/api-zod/src/generated/types/nodeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NodeResultOutput } from "./nodeResultOutput";
import type { NodeResultStatus } from "./nodeResultStatus";

export interface NodeResult {
  nodeId: string;
  nodeType: string;
  label: string;
  status: NodeResultStatus;
  output?: NodeResultOutput;
  reasoning?: string;
  startedAt?: Date;
  completedAt?: Date;
  durationMs?: number;
}